Refactored code, added system tray icon and hidden console.
