+++ mi/HidLibrary/DeviceMode.cs
@@ -1,0 +1,8 @@
+namespace HidLibrary
+{
+	public enum DeviceMode
+	{
+		NonOverlapped = 0,
+		Overlapped = 1
+	}
+}
