+++ mi/HidLibrary/ReadCallback.cs
@@ -1,0 +1,4 @@
+namespace HidLibrary
+{
+	public delegate void ReadCallback(HidDeviceData data);
+}
