+++ mi/HidLibrary/ReadReportCallback.cs
@@ -1,0 +1,4 @@
+namespace HidLibrary
+{
+	public delegate void ReadReportCallback(HidReport report);
+}
