+++ mi/HidLibrary/RemovedEventHandler.cs
@@ -1,0 +1,4 @@
+namespace HidLibrary
+{
+	public delegate void RemovedEventHandler();
+}
