+++ mi/HidLibrary/WriteCallback.cs
@@ -1,0 +1,4 @@
+namespace HidLibrary
+{
+	public delegate void WriteCallback(bool success);
+}
