+++ mi/ScpDriverInterface/X360Controller.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace ScpDriverInterface
+﻿namespace ScpDriverInterface
 {
 	/// <summary>
 	/// A virtual XBox 360 Controller. After setting the desired values, use the GetReport() method to generate a controller report that can be used with ScpBus's Report() method.
@@ -20,7 +18,7 @@ namespace ScpDriverInterface
 			RightStickX = 0;
 			RightStickY = 0;
 		}
-		
+
 		/// <summary>
 		/// Generates a new X360Controller object. Optionally, you can specify the initial state of the controller.
 		/// </summary>
@@ -123,35 +121,5 @@ namespace ScpDriverInterface
 
 			return bytes;
 		}
-	}
-
-	/// <summary>
-	/// The buttons to be used with an X360Controller object.
-	/// </summary>
-	[Flags]
-	public enum X360Buttons
-	{
-		None = 0,
-
-		Up = 1 << 0,
-		Down = 1 << 1,
-		Left = 1 << 2,
-		Right = 1 << 3,
-
-		Start = 1 << 4,
-		Back = 1 << 5,
-
-		LeftStick = 1 << 6,
-		RightStick = 1 << 7,
-
-		LeftBumper = 1 << 8,
-		RightBumper = 1 << 9,
-
-		Logo = 1 << 10,
-
-		A = 1 << 12,
-		B = 1 << 13,
-		X = 1 << 14,
-		Y = 1 << 15,
 	}
 }
