+++ mi/mi.csproj
@@ -5,7 +5,7 @@
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{6C084899-30AD-44EF-9975-699D75B5D61B}</ProjectGuid>
-    <OutputType>Exe</OutputType>
+    <OutputType>WinExe</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>mi</RootNamespace>
     <AssemblyName>mi</AssemblyName>
@@ -37,9 +37,17 @@
   <PropertyGroup>
     <ApplicationManifest>app.manifest</ApplicationManifest>
   </PropertyGroup>
+  <PropertyGroup>
+    <StartupObject />
+  </PropertyGroup>
+  <PropertyGroup>
+    <ApplicationIcon>Resources\micontroller.ico</ApplicationIcon>
+  </PropertyGroup>
   <ItemGroup>
     <Reference Include="System" />
     <Reference Include="System.Core" />
+    <Reference Include="System.Drawing" />
+    <Reference Include="System.Windows.Forms" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data.DataSetExtensions" />
     <Reference Include="Microsoft.CSharp" />
@@ -48,6 +56,14 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="AboutBox.cs">
+      <SubType>Form</SubType>
+    </Compile>
+    <Compile Include="AboutBox.Designer.cs">
+      <DependentUpon>AboutBox.cs</DependentUpon>
+    </Compile>
+    <Compile Include="ContextMenus.cs" />
+    <Compile Include="HidLibrary\DeviceMode.cs" />
     <Compile Include="HidLibrary\Extensions.cs" />
     <Compile Include="HidLibrary\HidAsyncState.cs" />
     <Compile Include="HidLibrary\HidDevice.cs" />
@@ -56,19 +72,47 @@
     <Compile Include="HidLibrary\HidDeviceData.cs" />
     <Compile Include="HidLibrary\HidDeviceEventMonitor.cs" />
     <Compile Include="HidLibrary\HidDevices.cs" />
+    <Compile Include="HidLibrary\HidEnumerator.cs" />
     <Compile Include="HidLibrary\HidFastReadDevice.cs" />
     <Compile Include="HidLibrary\HidReport.cs" />
     <Compile Include="HidLibrary\IHidDevice.cs" />
     <Compile Include="HidLibrary\IHidEnumerator.cs" />
+    <Compile Include="HidLibrary\InsertedEventHandler.cs" />
     <Compile Include="HidLibrary\NativeMethods.cs" />
+    <Compile Include="HidLibrary\ReadCallback.cs" />
+    <Compile Include="HidLibrary\ReadReportCallback.cs" />
+    <Compile Include="HidLibrary\RemovedEventHandler.cs" />
+    <Compile Include="HidLibrary\ShareMode.cs" />
+    <Compile Include="HidLibrary\WriteCallback.cs" />
+    <Compile Include="ProcessIcon.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Properties\Resources.Designer.cs">
+      <AutoGen>True</AutoGen>
+      <DesignTime>True</DesignTime>
+      <DependentUpon>Resources.resx</DependentUpon>
+    </Compile>
+    <Compile Include="ScpDriverInterface\NativeMethods.cs" />
     <Compile Include="ScpDriverInterface\ScpBus.cs" />
+    <Compile Include="ScpDriverInterface\X360Buttons.cs" />
     <Compile Include="ScpDriverInterface\X360Controller.cs" />
+    <Compile Include="Xiaomi_gamepad.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
     <None Include="app.manifest" />
+  </ItemGroup>
+  <ItemGroup>
+    <EmbeddedResource Include="AboutBox.resx">
+      <DependentUpon>AboutBox.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Properties\Resources.resx">
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>Resources.Designer.cs</LastGenOutput>
+    </EmbeddedResource>
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="Resources\micontroller.ico" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
